dashboard_app/requirements.txt
Flask==2.3.3
requests==2.31.0
bcrypt==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Session==0.5.0
Flask-Talisman==1.1.0
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Jinja2==3.1.2
pytest==7.4.3
pytest-cov==4.1.0
